third try.
Well, now it runs. But not complety. 
Cholesky is not running the way it should (a lot of 'nan' and 'inf') I'm not sure if that output is expected. Nonetheless, as soon as I run the code, the mean value turns to '-nan' . I'm still trying to figure that out.